fixes #421 (#430)
ty! next time can you try to remove 'line' changes your editor is making, cleans up the diff a little bit.. makes it easier to read. 

JetBrains and Sublime sometimes have a difference in opinions as to when "long lines" should be cut off. not that either is better or worse- just should be done as a sep PR to make things more concise and readable across changes..

ty!!!