checks ip mask to ensure only /32 (#498)
condition to only run this hunter if a single ipv4 is submitted, and add the `hunter` tag. an example problem scenario is if you submit a range like /22 or /18 to be allow listed. this hunter runs at the same time the router does on the zmq sub queue and creates a race condition. if the hunter adds first, the /24 will be created and the upsert on the router side will find the existing range added by the hunter and not add/upsert to the /22 that should have taken precedent.